Rename 'forte' module to 'class'
